requirements.txt
torch
torchvision
spaces
pillow
transformers>=4.39.1
pillow-heif